+++ dvh/dvh_bokeh_models/main/time_series.py
@@ -9,7 +9,7 @@ from bokeh.plotting import figure
 from bokeh.models import Select, TextInput, RadioGroup, Slider, Div, Legend, CustomJS, HoverTool, Button, Spacer
 from bokeh.layouts import column, row
 import options
-from utilities import clear_source_data, collapse_into_single_dates, moving_avg,\
+from tools.utilities import clear_source_data, collapse_into_single_dates, moving_avg,\
     moving_avg_by_calendar_day, clear_source_selection, get_csv
 from scipy.stats import ttest_ind, ranksums, normaltest
 import numpy as np
@@ -105,7 +105,7 @@ class TimeSeries:
 
         self.download_time_plot = Button(label="Download Plot Data", button_type="default", width=150)
         self.download_time_plot.callback = CustomJS(args=dict(source=self.sources.time_csv),
-                                                    code=open(join(dirname(__file__),
+                                                    code=open(join(dirname(dirname(__file__)),
                                                                    "download_new.js")).read())
 
         # histograms