+++ dvh/main.py
@@ -7,30 +7,33 @@ Created on Sun Apr 21 2017
 """
 
 from __future__ import print_function
-from utilities import load_options
-import auth
+import update_sys_path
+from tools.utilities import load_options
+from tools import auth
 from bokeh.layouts import row
 from bokeh.models import Spacer
 from bokeh.io import curdoc
 from bokeh.models.widgets import Button, Panel, Tabs, TextInput, PasswordInput
 import time
 import options
-import sources
-from custom_titles import custom_title
-from mlc_analyzer import MLC_Analyzer
-from time_series import TimeSeries
-from correlation import Correlation
-from roi_viewer import ROI_Viewer
-from rad_bio import RadBio
-from regression import Regression
-from query import Query
-from dvhs import DVHs
-from data_tables import DataTables
-from categories import Categories
-from planning_data import PlanningData
-from source_listener import SourceListener
+from dvh_bokeh_models.main.sources import Sources
+from dvh_bokeh_models.main.custom_titles import CustomTitles
+from dvh_bokeh_models.main.mlc_analyzer import MLC_Analyzer
+from dvh_bokeh_models.main.time_series import TimeSeries
+from dvh_bokeh_models.main.correlation import Correlation
+from dvh_bokeh_models.main.roi_viewer import ROI_Viewer
+from dvh_bokeh_models.main.rad_bio import RadBio
+from dvh_bokeh_models.main.regression import Regression
+from dvh_bokeh_models.main.query import Query
+from dvh_bokeh_models.main.dvhs import DVHs
+from dvh_bokeh_models.main.data_tables import DataTables
+from dvh_bokeh_models.main.categories import Categories
+from dvh_bokeh_models.main.planning_data import PlanningData
+from tools.source_listener import SourceListener
+
 
 options = load_options(options)
+custom_title = CustomTitles().custom_title
 
 
 # This depends on a user defined function in dvh/auth.py.  By default, this returns True
@@ -39,9 +42,13 @@ options = load_options(options)
 # Please see Bokeh documentation for more information
 ACCESS_GRANTED = not options.AUTH_USER_REQ
 
+
 # !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
 # Bokeh component classes
 # !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
+
+# ColumnDataSource objects
+sources = Sources()
 
 # Categories map of dropdown values, SQL column, and SQL table (and data source for range_categories)
 categories = Categories(sources)