+++ dvh/update_sys_path.py
@@ -1,0 +1,3 @@
+import sys
+from os.path import dirname, realpath
+sys.path.append(dirname(dirname(realpath(__file__))))